# lz4block: Use pointer-to-array in pool for up to 30% throughput improvement